src/requirements.txt
arrow==0.8.0
autoenv==1.0.0
Beaker==1.8.1
beautifulsoup4==4.5.1
binaryornot==0.4.0
boto==2.43.0
bottle==0.12.13
chardet==2.3.0
cheroot==5.2.0
CherryPy==10.1.1
click==6.6
cookiecutter==1.4.0
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.6
django-allauth==0.27.0
django-angular==0.8.4
django-appconf==1.0.2
django-channels==0.7.0
django-chartjs==1.1
django-classy-tags==0.6.2
django-compressor==2.1
django-cookie-law==1.0.11
django-crispy-forms==1.6.0
django-easy-maps==0.9.2
django-embed-video==1.1.0
django-emoji==2.2.0
django-environ==0.4.0
django-extensions==1.7.4
django-form-utils==1.0.3
django-haystack==2.5.0
django-hitcount==1.2.2
django-jquery==3.1.0
django-markdown-deux==1.0.5
django-modelcluster==2.0
django-mptt==0.8.6
django-pagedown==0.1.1
django-qrcode==0.3
django-registration-redux==1.4
django-sitetree==1.6.0
django-static-jquery==2.1.4
django-storages==1.5.1
django-tagging==0.4.5
django-taggit==0.18.3
django-treebeard==4.0.1
djangorestframework==3.6.2
first==2.0.1
Flask==0.11.1
freeze==1.0.10
future==0.15.2
geopy==1.11.0
gunicorn==19.6.0
html5lib==1.0b10
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.8
jinja2-time==0.2.0
Mako==1.0.6
Markdown==2.6.6
markdown2==2.3.3
MarkupSafe==0.23
numpy==1.12.0
oauthlib==2.0.0
pandas==0.19.2
Pillow==3.3.1
pip-tools==1.8.0
portend==1.8
poyo==0.4.0
psycopg2==2.6.2
PyMySQL==0.7.9
PyPDF2==1.26.0
pypiwin32==219
python-dateutil==2.5.3
python3-openid==3.0.10
pytz==2016.6.1
PyYAML==3.12
qrcode==5.3
rcssmin==1.0.6
reportlab==3.3.0
requests==2.11.1
requests-oauthlib==0.6.2
rest-pandas==0.5.0
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.3
South==1.0.2
tempora==1.6.1
Unidecode==0.4.19
wagtail==1.6.2
webencodings==0.5
Werkzeug==0.11.11
whichcraft==0.4.0
whitenoise==3.2.1
Willow==0.3.1
xhtml2pdf==0.2b1
